Introduce LiveBytesStats. Cleanup
